--- basic/src/main/java/com/pgms/part1/VoucherApplication.java
@@ -1,13 +1,0 @@
-package com.pgms.part1;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class VoucherApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(VoucherApplication.class, args);
-	}
-
-}
