--- basic/src/main/java/com/pgms/part1/domain/customer/dto/CustomerResponseDto.java
@@ -1,6 +1,0 @@
-package com.pgms.part1.domain.customer.dto;
-
-import java.util.UUID;
-
-public record CustomerResponseDto(UUID id, Boolean isBlocked) {
-}
