--- basic/src/main/java/com/pgms/part1/domain/customer/repository/CustomerRepository.java
@@ -1,9 +1,0 @@
-package com.pgms.part1.domain.customer.repository;
-
-import com.pgms.part1.domain.customer.dto.CustomerResponseDto;
-
-import java.util.List;
-
-public interface CustomerRepository {
-    public List<CustomerResponseDto> listBlockedCustomers();
-}
