--- basic/src/main/java/com/pgms/part1/domain/voucher/dto/VoucherCreateRequestDto.java
@@ -1,4 +1,0 @@
-package com.pgms.part1.domain.voucher.dto;
-
-public record VoucherCreateRequestDto(Integer command, Integer discount) {
-}
