--- basic/src/main/java/com/pgms/part1/domain/voucher/dto/VoucherMenuRequestDto.java
@@ -1,3 +1,0 @@
-package com.pgms.part1.domain.voucher.dto;
-
-public record VoucherMenuRequestDto(String command){}
