+++ src/main/java/com/pgms/part1/domain/customer/dto/CustomerResponseDto.java
@@ -1,0 +1,6 @@
+package java.com.pgms.part1.domain.customer.dto;
+
+import java.util.UUID;
+
+public record CustomerResponseDto(UUID id, Boolean isBlocked) {
+}
