+++ src/main/java/com/pgms/part1/domain/customer/service/CustomerService.java
@@ -1,0 +1,20 @@
+package java.com.pgms.part1.domain.customer.service;
+
+import com.pgms.part1.domain.customer.dto.CustomerResponseDto;
+import com.pgms.part1.domain.customer.repository.CustomerRepository;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class CustomerService {
+    private final CustomerRepository customerRepository;
+
+    public CustomerService(CustomerRepository customerRepository) {
+        this.customerRepository = customerRepository;
+    }
+
+    public List<CustomerResponseDto> listBlockedCustomers(){
+        return customerRepository.listBlockedCustomers();
+    }
+}
