+++ src/main/java/com/pgms/part1/domain/voucher/dto/VoucherMenuRequestDto.java
@@ -1,0 +1,3 @@
+package java.com.pgms.part1.domain.voucher.dto;
+
+public record VoucherMenuRequestDto(String command){}
