+++ src/main/java/com/pgms/part1/domain/voucher/entity/PercentDiscountVoucher.java
@@ -1,0 +1,9 @@
+package java.com.pgms.part1.domain.voucher.entity;
+
+import java.util.UUID;
+
+public class PercentDiscountVoucher extends Voucher {
+    public PercentDiscountVoucher(UUID id, Integer discount) {
+        super(id, discount, VoucherDiscountType.PERCENT_DISCOUNT);
+    }
+}
