+++ src/main/java/com/pgms/part1/domain/voucher/repository/VoucherRepository.java
@@ -1,0 +1,12 @@
+package java.com.pgms.part1.domain.voucher.repository;
+
+import com.pgms.part1.domain.voucher.entity.Voucher;
+
+import java.util.List;
+
+public interface VoucherRepository {
+
+    public List<Voucher> list();
+
+    public void add(Voucher voucher);
+}
